+++ Engine/CinderEngine/include/Engine/System/Events/EventSystem/EventSystem.h
@@ -29,7 +29,7 @@ namespace Framework
     EventSystem();
     ~EventSystem();
 
-    virtual void Update(const double& dt);
+    virtual void Update(const float& dt);
 
     virtual const string GetName() { return "EventSystem"; }
